[modified] Renamed app messagememos to 'feedbacks'
