+++ feedbacks/urls.py
@@ -2,7 +2,7 @@ from django.urls import path
 from . import views
 
 
-app_name = "messagememos"
+app_name = "feedbacks"
 
 urlpatterns = [
     path("", views.create, name="create"),
