+++ templates/accounts/user.html
@@ -86,7 +86,7 @@
       <div class="">
         <div class="flex justify-evenly items-start bg-white rounded-lg p-4 shadow">
           <form
-            action="{% url 'messagememos:create' %}"
+            action="{% url 'feedbacks:create' %}"
             method="post"
             class="pt-10"
           >
